refactor: Member, Point 도메인 테스트 추가
