+++ server/src/test/java/com/example/tyfserver/banner/service/BannerServiceTest.java
@@ -4,7 +4,7 @@ import com.example.tyfserver.auth.dto.LoginMember;
 import com.example.tyfserver.banner.domain.Banner;
 import com.example.tyfserver.banner.domain.BannerRepository;
 import com.example.tyfserver.banner.dto.BannerResponse;
-import com.example.tyfserver.member.MemberTest;
+import com.example.tyfserver.member.domain.MemberTest;
 import com.example.tyfserver.member.domain.Member;
 import com.example.tyfserver.member.repository.MemberRepository;
 import org.junit.jupiter.api.AfterEach;
