+++ server/src/test/java/com/example/tyfserver/member/domain/MemberTest.java
@@ -1,0 +1,32 @@
+package com.example.tyfserver.member.domain;
+
+import com.example.tyfserver.auth.domain.Oauth2Type;
+import com.example.tyfserver.donation.domain.Donation;
+import com.example.tyfserver.donation.domain.Message;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class MemberTest {
+
+    public static Member testMember() {
+        return new Member("tyf@gmail.com", "nickname", "pageName", Oauth2Type.NAVER);
+    }
+
+    public static Member testMember2() {
+        return new Member("tyf2@gmail.com", "nickname2", "pageName2", Oauth2Type.NAVER);
+    }
+
+    @Test
+    @DisplayName("addDonation 메서드 테스트")
+    public void addDonationTest() {
+        //given
+        Member member = testMember();
+        Donation donation = new Donation(1000L, new Message("name", "message", false));
+        //when
+        member.addDonation(donation);
+        //then
+        assertThat(member.getPoint()).isEqualTo(1000L);
+    }
+}
